GeometryProcessing: Mesh Solvers.   Fix implicit cast warnings.  These are related the integer size type used by Eigen, and our comparatively  modest maxtrix  row/col max of Max_int32.
#preflight 62bb18e31813601bbecdfc25
#rb Jimmy.Andrews

[CL 20971713 by David Hill in ue5-main branch]

+++ Engine/Plugins/Runtime/GeometryProcessing/Source/DynamicMesh/Private/Solvers/Internal/LaplacianOperators.cpp
@@ -138,7 +138,7 @@ double ConstructScaledCotangentLaplacian(const FDynamicMesh3& DynamicMesh, FVert
 	ConstructCotangentLaplacian(DynamicMesh, VertexMap, AreaMatrix, CotangentInterior, CotangentBoundary);
 
 	// Find average entry in the area matrix
-	const int32 Rank = AreaMatrix.cols();
+	const int32 Rank = (int32)AreaMatrix.cols();
 	double AveArea = 0.;
 	for (int32 i = 0; i < Rank; ++i)
 	{
